Friday changes

+++ clock.h
@@ -1,14 +1,13 @@
 #ifndef CLOCK_H
 #define CLOCK_H
 
-#include <stdint.h>             // uint16_t
 #include <stdio.h>              // printf
 #include <string>               // std::string
 
 class Clock {
 private:
     static Clock* clk_instance; // Singleton Object
-    uint16_t cycle;             // Current cycle of clock
+    int cycle;                  // Current cycle of clock
     Clock();                    // Forces Public Instantiation
 public:
     // Singleton Method
